+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/actionSystem/CCv2GroovyReplicaSelectionModeAction.kt
@@ -26,9 +26,7 @@ import sap.commerce.toolset.Plugin
 import sap.commerce.toolset.ccv2.CCv2ExecConstants
 import sap.commerce.toolset.ccv2.ui.CCv2ReplicaSelectionDialog
 import sap.commerce.toolset.groovy.actionSystem.GroovyReplicaSelectionModeAction
-import sap.commerce.toolset.groovy.editor.groovyExecContextSettings
-import sap.commerce.toolset.groovy.exec.context.GroovyExecContext
-import sap.commerce.toolset.hac.exec.HacExecConnectionService
+import sap.commerce.toolset.groovy.exec.GroovyExecService
 import java.awt.Component
 
 class CCv2GroovyReplicaSelectionModeAction : GroovyReplicaSelectionModeAction(CCv2ExecConstants.ccv2) {
@@ -42,10 +40,7 @@ class CCv2GroovyReplicaSelectionModeAction : GroovyReplicaSelectionModeAction(CC
         val project = e.project ?: return
         val virtualFile = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
         val component = e.inputEvent?.source?.asSafely<Component>() ?: return
-        val execSettings = e.groovyExecContextSettings {
-            val activeConnection = HacExecConnectionService.getInstance(project).activeConnection
-            GroovyExecContext.defaultSettings(activeConnection)
-        }
+        val execSettings = GroovyExecService.getInstance(project).getSettings(virtualFile)
 
         val dialog = CCv2ReplicaSelectionDialog(project, virtualFile, execSettings, component)
         dialog.showAndGet()