+++ modules/groovy/exec/src/sap/commerce/toolset/groovy/exec/GroovyExecService.kt
@@ -22,18 +22,24 @@ import com.intellij.openapi.application.edtWriteAction
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiDocumentManager
 import kotlinx.coroutines.launch
 import kotlinx.serialization.json.Json
 import sap.commerce.toolset.Notifications
 import sap.commerce.toolset.groovy.exec.context.GroovyExecContext
+import sap.commerce.toolset.groovy.exec.context.GroovyExecContext.Settings
+import sap.commerce.toolset.groovy.exec.context.GroovyReplicaAwareContext
+import sap.commerce.toolset.groovy.getSpringContextMode
 import sap.commerce.toolset.groovy.groovyRemoteSpringBeans
+import sap.commerce.toolset.groovy.groovySettings
 import sap.commerce.toolset.groovy.lang.resolve.RemoteSpringBean
 import sap.commerce.toolset.hac.exec.HacExecConnectionService
 import sap.commerce.toolset.readResource
 import sap.commerce.toolset.settings.state.SpringContextMode
 import sap.commerce.toolset.settings.state.TransactionMode
+import sap.commerce.toolset.settings.yDeveloperSettings
 
 @Service(Service.Level.PROJECT)
 class GroovyExecService(private val project: Project) {
@@ -41,7 +47,7 @@ class GroovyExecService(private val project: Project) {
     private var fetching = false
 
     fun fetchRemoteSpringBeans(virtualFile: VirtualFile) {
-        val webContext = virtualFile.groovyWebContext
+        val webContext = getActiveWebContext(virtualFile)
 
         val server = HacExecConnectionService.getInstance(project).activeConnection
         val groovyScript = readResource("scripts/groovy-fetchSpringBeans.groovy")
@@ -99,7 +105,8 @@ class GroovyExecService(private val project: Project) {
                         """
                             Web context: $webContext<br>
                             Beans: ${beans.size}
-                        """.trimIndent())
+                        """.trimIndent()
+                    )
                     .notify(project)
             }
         }
@@ -119,12 +126,12 @@ class GroovyExecService(private val project: Project) {
         GroovyExecClient.getInstance(project).execute(
             context,
             beforeCallback = {
-                virtualFile.groovyWebContextsFetching = true
+                setFetchingWebContexts(virtualFile, true)
                 virtualFile.groovyExecContextSettings = virtualFile.groovyExecContextSettings
                     ?.copy(webContext = null)
             },
             onError = { _, ex ->
-                virtualFile.groovyWebContextsFetching = false
+                setFetchingWebContexts(virtualFile, false)
                 Notifications
                     .error("Unable to load web contexts", ex.message ?: "")
                     .notify(project)
@@ -135,8 +142,8 @@ class GroovyExecService(private val project: Project) {
                 ?.filter { it.isNotBlank() }
                 ?.sorted()
 
-            virtualFile.groovyWebContextsFetching = false
-            virtualFile.groovyWebContexts = contexts
+            setFetchingWebContexts(virtualFile, false)
+            setWebContexts(virtualFile, contexts)
 
             Notifications
                 .info("Found ${contexts?.size ?: 0} web contexts")
@@ -144,22 +151,69 @@ class GroovyExecService(private val project: Project) {
         }
     }
 
-    fun activateWebContext(virtualFile: VirtualFile, webContext: String) {
+    fun initSettings(virtualFile: VirtualFile): Settings {
+        val activeConnection = HacExecConnectionService.getInstance(project).activeConnection
+        val groovySettings = project.yDeveloperSettings.groovySettings
+
+        return GroovyExecContext.defaultSettings(activeConnection, groovySettings).also {
+            virtualFile.groovyExecContextSettings = it
+        }
+    }
+
+    fun getSettings(virtualFile: VirtualFile): Settings = virtualFile.groovyExecContextSettings
+        ?: initSettings(virtualFile)
+
+    fun setSettings(virtualFile: VirtualFile, settings: Settings) {
+        virtualFile.groovyExecContextSettings = settings
+    }
+
+    fun getTransactionMode(virtualFile: VirtualFile, project: Project) = virtualFile.groovyExecContextSettings
+        ?.transactionMode
+        ?: project.yDeveloperSettings.groovySettings.transactionMode
+
+    fun setTransactionMode(virtualFile: VirtualFile, mode: TransactionMode) {
+        virtualFile.groovyExecContextSettings = getSettings(virtualFile).copy(transactionMode = mode)
+    }
+
+    fun getReplicaAwareContext(virtualFile: VirtualFile) = virtualFile.groovyExecContextSettings?.replicaContext
+        ?: GroovyReplicaAwareContext.auto()
+
+    fun setReplicaAwareContext(virtualFile: VirtualFile, context: GroovyReplicaAwareContext) {
+        virtualFile.groovyExecContextSettings = getSettings(virtualFile).copy(replicaContext = context)
+    }
+
+    fun getWebContexts(virtualFile: VirtualFile) = virtualFile.getUserData(KEY_WEB_CONTEXTS)
+    fun setWebContexts(virtualFile: VirtualFile, webContexts: Collection<String>?) = virtualFile.putUserData(KEY_WEB_CONTEXTS, webContexts)
+
+    fun isFetchingWebContexts(virtualFile: VirtualFile) = virtualFile.getUserData(KEY_WEB_CONTEXTS_FETCHING) ?: false
+    private fun setFetchingWebContexts(virtualFile: VirtualFile, state: Boolean) = virtualFile.putUserData(KEY_WEB_CONTEXTS_FETCHING, state)
+
+    fun getActiveWebContext(virtualFile: VirtualFile) = virtualFile.groovyExecContextSettings?.webContext
+        ?: GroovyExecConstants.DEFAULT_WEB_CONTEXT
+
+    fun activateWebContext(project: Project, virtualFile: VirtualFile, webContext: String) {
         val activeWebContext = if (webContext != GroovyExecConstants.DEFAULT_WEB_CONTEXT) webContext
         else null
 
-        virtualFile.groovyExecContextSettings = virtualFile.groovyExecContextSettings {
-            val activeConnection = HacExecConnectionService.getInstance(project).activeConnection
-            GroovyExecContext.defaultSettings(activeConnection)
-        }.copy(webContext = activeWebContext)
+        virtualFile.groovyExecContextSettings = getSettings(virtualFile).copy(webContext = activeWebContext)
 
-        val currentSpringContextMode = virtualFile.getCurrentSpringContextMode(project)
+        val currentSpringContextMode = virtualFile.getSpringContextMode(project)
         if (currentSpringContextMode == SpringContextMode.REMOTE) {
             fetchRemoteSpringBeans(virtualFile)
         }
     }
 
+    private var VirtualFile.groovyExecContextSettings
+        get() = this.getUserData(KEY_EXECUTION_SETTINGS)
+        set(value) {
+            this.putUserData(KEY_EXECUTION_SETTINGS, value)
+        }
+
     companion object {
+        private val KEY_EXECUTION_SETTINGS = Key.create<Settings>("sap.cx.groovy.execution.settings")
+        private val KEY_WEB_CONTEXTS = Key.create<Collection<String>>("sap.cx.groovy.execution.webContexts")
+        private val KEY_WEB_CONTEXTS_FETCHING = Key.create<Boolean>("sap.cx.groovy.execution.webContexts.fetching")
+
         fun getInstance(project: Project): GroovyExecService = project.service()
     }
 }