+++ modules/groovy/exec/src/sap/commerce/toolset/groovy/exec/context/GroovyExecContext.kt
@@ -21,9 +21,9 @@ package sap.commerce.toolset.groovy.exec.context
 import org.apache.commons.lang3.BooleanUtils
 import sap.commerce.toolset.exec.context.ExecContext
 import sap.commerce.toolset.exec.context.ReplicaContext
-import sap.commerce.toolset.groovy.exec.GroovyExecExceptionHandling
-import sap.commerce.toolset.groovy.exec.GroovyExecMode
-import sap.commerce.toolset.hac.HacExecConstants
+import sap.commerce.toolset.groovy.settings.GroovyDeveloperSettings
+import sap.commerce.toolset.groovy.settings.state.GroovyExecExceptionHandling
+import sap.commerce.toolset.groovy.settings.state.GroovyExecMode
 import sap.commerce.toolset.hac.exec.settings.state.HacConnectionSettingsState
 import sap.commerce.toolset.settings.state.TransactionMode
 
@@ -36,7 +36,7 @@ data class GroovyExecContext(
     val transactionMode: TransactionMode,
     val webContext: String? = null,
     val replicaContext: ReplicaContext? = null,
-    val executionMode: GroovyExecMode = GroovyExecMode.DIRECT
+    val execMode: GroovyExecMode = GroovyExecMode.DIRECT
 ) : ExecContext {
 
     constructor(
@@ -51,6 +51,7 @@ data class GroovyExecContext(
         content = content,
         timeout = settings.timeout,
         exceptionHandling = settings.exceptionHandling,
+        execMode = settings.execMode,
         transactionMode = settings.transactionMode,
         webContext = settings.webContext,
         replicaContext = replicaContext,
@@ -64,9 +65,10 @@ data class GroovyExecContext(
 
     data class Settings(
         override val timeout: Int,
-        val exceptionHandling: GroovyExecExceptionHandling,
         val webContext: String? = null,
-        val transactionMode: TransactionMode = TransactionMode.ROLLBACK,
+        val exceptionHandling: GroovyExecExceptionHandling,
+        val execMode: GroovyExecMode,
+        val transactionMode: TransactionMode,
         val replicaContext: GroovyReplicaAwareContext = GroovyReplicaAwareContext.auto()
     ) : ExecContext.Settings {
         override fun mutable() = Mutable(
@@ -75,11 +77,13 @@ data class GroovyExecContext(
             replicaContext = replicaContext,
             webContext = webContext,
             exceptionHandling = exceptionHandling,
+            execMode = this@Settings.execMode,
         )
 
         data class Mutable(
             override var timeout: Int,
             var exceptionHandling: GroovyExecExceptionHandling,
+            var execMode: GroovyExecMode,
             var transactionMode: TransactionMode,
             var replicaContext: GroovyReplicaAwareContext,
             var webContext: String?
@@ -90,6 +94,7 @@ data class GroovyExecContext(
                 replicaContext = replicaContext,
                 webContext = webContext,
                 exceptionHandling = exceptionHandling,
+                execMode = execMode,
             )
         }
     }
@@ -97,10 +102,11 @@ data class GroovyExecContext(
     companion object {
         const val DEFAULT_TITLE = "Executing Groovy script on the remote SAP Commerce instance..."
 
-        fun defaultSettings(connectionSettings: HacConnectionSettingsState? = null) = Settings(
-            timeout = connectionSettings?.timeout ?: HacExecConstants.DEFAULT_TIMEOUT,
-            transactionMode = TransactionMode.ROLLBACK,
-            exceptionHandling = GroovyExecExceptionHandling.FULL_STACKTRACE,
+        fun defaultSettings(connectionSettings: HacConnectionSettingsState, groovySettings: GroovyDeveloperSettings) = Settings(
+            timeout = connectionSettings.timeout,
+            transactionMode = groovySettings.transactionMode,
+            execMode = groovySettings.execMode,
+            exceptionHandling = groovySettings.exceptionHandling,
         )
     }
 }