+++ modules/groovy/ui/src/sap/commerce/toolset/groovy/editor/GroovySplitEditor.kt
@@ -19,7 +19,6 @@
 package sap.commerce.toolset.groovy.editor
 
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.application.edtWriteAction
 import com.intellij.openapi.fileEditor.FileEditor
@@ -36,8 +35,9 @@ import com.intellij.ui.OnePixelSplitter
 import com.intellij.util.asSafely
 import kotlinx.coroutines.launch
 import sap.commerce.toolset.exec.context.DefaultExecResult
-import sap.commerce.toolset.groovy.exec.context.GroovyExecContext
-import sap.commerce.toolset.groovy.exec.groovyExecContextSettings
+import sap.commerce.toolset.groovy.exec.GroovyExecService
+import sap.commerce.toolset.groovy.getSpringContextMode
+import sap.commerce.toolset.groovy.setSpringContextMode
 import java.awt.BorderLayout
 import java.beans.PropertyChangeListener
 import java.io.Serial
@@ -46,10 +46,6 @@ import javax.swing.JPanel
 
 fun AnActionEvent.groovySplitEditor() = this.getData(PlatformDataKeys.FILE_EDITOR)
     ?.asSafely<GroovySplitEditor>()
-
-fun AnActionEvent.groovyExecContextSettings(fallback: () -> GroovyExecContext.Settings) = this.getData(CommonDataKeys.VIRTUAL_FILE)
-    ?.groovyExecContextSettings(fallback)
-    ?: fallback()
 
 class GroovySplitEditor(internal val textEditor: TextEditor, private val project: Project) : UserDataHolderBase(), FileEditor, TextEditor {
 
@@ -83,6 +79,14 @@ class GroovySplitEditor(internal val textEditor: TextEditor, private val project
 
     private val rootPanel = JPanel(BorderLayout()).apply {
         add(verticalSplitter, BorderLayout.CENTER)
+    }
+
+    init {
+        textEditor.editor.virtualFile
+            ?.let {
+                it.setSpringContextMode(it.getSpringContextMode(project))
+                GroovyExecService.getInstance(project).initSettings(it)
+            }
     }
 
     fun renderExecutionResults(results: Collection<DefaultExecResult>) = GroovyInEditorResultsView.getInstance(project).resultView(this, results) { coroutineScope, view ->